# api/openapi/src/catalog.yaml
openapi: 3.0.3
info:
  title: Catalog REST API
  description: REST API for Kubeflow Hub Catalog
  version: v1alpha1
  license:
    name: Apache 2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0"
servers:
  - url: "https://localhost:8080"
  - url: "http://localhost:8080"
paths:
...
  /api/mcp_catalog/v1alpha1/mcp_servers:
    description: >-
      The REST endpoint/path used to list zero or more `MCPServer` entities.
    get:
      summary: List MCP servers.
      tags:
        - MCPCatalogService
      parameters:
        - name: name
          description: Filter MCP servers by server name using SQL LIKE pattern matching.
          schema:
            type: string
          in: query
          required: false
        - name: q
          description: Free-form keyword search across name, description, and provider.
          schema:
            type: string
          in: query
          required: false
        - name: sourceLabel
          description: |-
            Filter MCP servers by the label associated with the source. Multiple
            values can be separated by commas. If one of the values is the
            string `null`, then MCP servers from every source without a label will
            be returned.
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
        - $ref: "#/components/parameters/mcpServerFilterQuery"
        - $ref: "#/components/parameters/namedQuery"
        - $ref: "#/components/parameters/includeTools"
        - $ref: "#/components/parameters/toolLimit"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/orderBy"
        - $ref: "#/components/parameters/sortOrder"
        - $ref: "#/components/parameters/nextPageToken"
      responses:
        "200":
          $ref: "#/components/responses/MCPServerListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: findMCPServers
  /api/mcp_catalog/v1alpha1/mcp_servers/filter_options:
    description: Lists options for `filterQuery` when listing MCP servers.
    get:
      summary: Lists fields, values, and named queries that can be used in `filterQuery` on the list MCP servers endpoint.
      tags:
        - MCPCatalogService
      responses:
        "200":
          $ref: "#/components/responses/FilterOptionsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: findMCPServersFilterOptions
  /api/mcp_catalog/v1alpha1/mcp_servers/{server_id}:
    description: >-
      The REST endpoint/path used to get an `MCPServer`.
    get:
      summary: Get an `MCPServer`.
      tags:
        - MCPCatalogService
      responses:
        "200":
          $ref: "#/components/responses/MCPServerResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: getMCPServer
    parameters:
      - name: server_id
        description: A unique identifier for an `MCPServer`.
        schema:
          type: string
        in: path
        required: true
      - $ref: "#/components/parameters/includeTools"
      - $ref: "#/components/parameters/toolLimit"
  /api/mcp_catalog/v1alpha1/mcp_servers/{server_id}/tools:
    description: >-
      The REST endpoint/path used to list zero or more `MCPTool` entities for an `MCPServer`.
    get:
      summary: List tools exposed by an `MCPServer`.
      tags:
        - MCPCatalogService
      parameters:
        - $ref: "#/components/parameters/mcpToolFilterQuery"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/orderBy"
        - $ref: "#/components/parameters/sortOrder"
        - $ref: "#/components/parameters/nextPageToken"
      responses:
        "200":
          $ref: "#/components/responses/MCPToolsListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: findMCPServerTools
    parameters:
      - name: server_id
        description: A unique identifier for an `MCPServer`.
        schema:
          type: string
        in: path
        required: true
  /api/mcp_catalog/v1alpha1/mcp_servers/{server_id}/tools/{tool_name}:
    description: >-
      The REST endpoint/path used to get a specific `MCPTool` from an `MCPServer`.
    get:
      summary: Get an `MCPTool` from an `MCPServer`.
      tags:
        - MCPCatalogService
      responses:
        "200":
          $ref: "#/components/responses/MCPToolWithServerResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: getMCPServerTool
    parameters:
      - name: server_id
        description: A unique identifier for an `MCPServer`.
        schema:
          type: string
        in: path
        required: true
      - name: tool_name
        description: A unique identifier for an `MCPTool` within the MCP server.
        schema:
          type: string
        in: path
        required: true
...
components:
  schemas:
...
    MCPServer:
      description: An MCP server in the model catalog.
      allOf:
        - $ref: "#/components/schemas/BaseResource"
        - type: object
          required:
            - name
            - toolCount
          properties:
            name:
              type: string
              description: Human-readable MCP server name. Must be unique within a source.
            source_id:
              type: string
              description: Catalog source that provides this server.
            provider:
              type: string
              description: Organization providing the server.
            logo:
              type: string
              description: MCP server logo.
            version:
              type: string
              description: Semantic version string of the MCP server.
            tags:
              type: array
              description: Categorization tags for this MCP server.
              example:
                - observability
                - monitoring
                - apm
              items:
                type: string
            license:
              type: string
              description: SPDX identifier for the server license.
            license_link:
              type: string
              format: uri
              description: URL to the full license text.
            toolCount:
              type: integer
              description: Number of tools exposed by this MCP server.
              example: 15
            tools:
              type: array
              description: Array of tools exposed by this MCP server.
              items:
                $ref: "#/components/schemas/MCPTool"
            securityIndicators:
              $ref: "#/components/schemas/MCPSecurityIndicator"
            deploymentMode:
              type: string
              description: Deployment mode for the MCP server.
              enum:
                - local
                - remote
            transports:
              type: array
              description: Supported transport protocols.
              items:
                type: string
                enum:
                  - stdio
                  - http
                  - sse
            artifacts:
              type: array
              description: OCI artifacts used for local deployment.
              items:
                $ref: "#/components/schemas/MCPArtifact"
            endpoints:
              $ref: "#/components/schemas/MCPEndpoints"
            readme:
              type: string
              description: Full Markdown documentation for this MCP server.
            documentationUrl:
              type: string
              format: uri
              description: URL to external documentation.
            repositoryUrl:
              type: string
              format: uri
              description: URL to source repository.
            sourceCode:
              type: string
              format: uri
              description: URL to browsable source code.
            publishedDate:
              type: string
              format: date-time
              description: Initial publication timestamp for the server.
            lastUpdated:
              type: string
              format: date-time
              description: Last update timestamp for the server metadata.
            runtimeMetadata:
              $ref: "#/components/schemas/MCPRuntimeMetadata"
              description: |-
                Runtime metadata for deploying this MCP server via MCPServer CRDs.
                Provides defaults, recommendations, and requirements for creating deployments.
                This is discovery metadata, not actual deployment configuration.
    MCPServerList:
      description: List of MCP server entities.
      allOf:
        - type: object
          properties:
            items:
              description: Array of `MCPServer` entities.
              type: array
              items:
                $ref: "#/components/schemas/MCPServer"
          required:
            - items
        - $ref: "#/components/schemas/BaseResourceList"
    MCPSecurityIndicator:
      type: object
      description: Security posture metadata for an MCP server.
      properties:
        verifiedSource:
          type: boolean
          description: Whether the source of this server has been verified.
        secureEndpoint:
          type: boolean
          description: Whether the server exposes secure communication endpoints.
        sast:
          type: boolean
          description: Whether static analysis has been performed.
        readOnlyTools:
          type: boolean
          description: Whether all exposed tools are read-only.
    MCPArtifact:
      description: Artifact metadata for local MCP server deployment.
      allOf:
        - $ref: "#/components/schemas/BaseResourceDates"
        - type: object
          required:
            - uri
          properties:
            uri:
              type: string
              format: uri
              description: OCI image URI for local deployment.
    MCPEndpoints:
      type: object
      description: Remote endpoint metadata for MCP server access.
      properties:
        http:
          type: string
          format: uri
          description: HTTP endpoint URL for the MCP server.
        sse:
          type: string
          format: uri
          description: SSE endpoint URL for the MCP server.
    MCPTool:
      description: An MCP tool exposed by an MCP server.
      allOf:
        - $ref: "#/components/schemas/BaseResource"
        - type: object
          required:
            - name
            - accessType
          properties:
            name:
              type: string
              description: Tool identifier.
            accessType:
              type: string
              description: Access mode supported by this tool.
              enum:
                - read_only
                - read_write
                - execute
            parameters:
              type: array
              description: Input parameters accepted by this tool.
              items:
                $ref: "#/components/schemas/MCPToolParameter"
    MCPToolWithServer:
      type: object
      required:
        - serverId
        - serverName
        - tool
      properties:
        serverId:
          type: string
          description: ID of the MCP server that exposes this tool
          example: prometheus-mcp
        serverName:
          type: string
          description: Human-readable name of the MCP server
          example: Prometheus MCP Server
        tool:
          $ref: "#/components/schemas/MCPTool"
    MCPToolParameter:
      type: object
      description: Metadata describing a single MCP tool parameter.
      required:
        - name
        - type
        - required
      properties:
        name:
          type: string
          description: Parameter name.
        type:
          type: string
          description: Parameter type.
        required:
          type: boolean
          description: Whether this parameter is required.
        description:
          type: string
          description: Parameter purpose.
    MCPToolsList:
      description: List of MCP tool entities.
      allOf:
        - type: object
          properties:
            items:
              description: Array of `MCPTool` entities.
              type: array
              items:
                $ref: "#/components/schemas/MCPTool"
          required:
            - items
        - $ref: "#/components/schemas/BaseResourceList"
    MCPRuntimeMetadata:
      type: object
      properties:
        defaultPort:
          type: integer
          description: |-
            Default port the MCP server listens on. Maps to MCPServer.spec.mcpPort.
            Should align with the primary transport type.
          example: 8080
          minimum: 1
          maximum: 65535
        defaultArgs:
          type: array
          description: |-
            Default command-line arguments for the MCP server.
            Users can override or extend these in the MCPServer CRD.
          items:
            type: string
          example: ["--log-level", "info", "--metrics-enabled"]
        requiredEnvironmentVariables:
          type: array
          description: |-
            Environment variables that MUST be set for the server to function.
            Documents the names and purposes - values come from user's Secrets/ConfigMaps.
          items:
            $ref: "#/components/schemas/MCPEnvVarMetadata"
        optionalEnvironmentVariables:
          type: array
          description: |-
            Optional environment variables that configure server behavior.
            Documents the names, purposes, and default values.
          items:
            $ref: "#/components/schemas/MCPEnvVarMetadata"
        recommendedResources:
          $ref: "#/components/schemas/MCPResourceRecommendation"
          description: |-
            Recommended CPU and memory resources based on typical workloads.
            Users should adjust based on their specific usage patterns.
        healthEndpoints:
          type: object
          description: Health check endpoint paths
          properties:
            liveness:
              type: string
              description: Liveness probe endpoint path
              example: /health
            readiness:
              type: string
              description: Readiness probe endpoint path
              example: /ready
        capabilities:
          type: object
          description: |-
            Runtime capabilities and requirements for this MCP server.
            Helps users understand what the server needs to run properly.
          properties:
            requiresNetwork:
              type: boolean
              description: Whether server requires network access
              default: true
            requiresFileSystem:
              type: boolean
              description: Whether server requires writable filesystem
              default: false
            requiresGPU:
              type: boolean
              description: Whether server requires GPU access
              default: false
        mcpPath:
          type: string
          description: |-
            HTTP path where MCP server accepts requests.
            Used for HTTP and SSE transports.
            Aligns with MCP Lifecycle operator conventions.
          pattern: '^/[a-zA-Z0-9/_.-]*$'
          default: /mcp
          example: /mcp
        prerequisites:
          $ref: "#/components/schemas/MCPPrerequisites"
          description: |-
            Kubernetes prerequisites discovery metadata.
            Helps users understand what K8s resources to create before deployment.
    MCPEnvVarMetadata:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          description: Environment variable name
          example: API_KEY
        description:
          type: string
          description: Purpose and usage of this variable
          example: API key for authenticating with external service
        required:
          type: boolean
          description: Whether this variable must be set
          default: false
        defaultValue:
          type: string
          description: |-
            Default value if not provided. Only for optional variables.
            Secrets should NEVER have default values.
          example: info
        type:
          type: string
          enum: [string, int, boolean, secret, config]
          description: Expected variable type
          default: string
        example:
          type: string
          description: Example value for documentation purposes
          example: my-api-key-123
    MCPResourceRecommendation:
      type: object
      properties:
        minimal:
          type: object
          description: Minimum viable resources for low-traffic scenarios
          properties:
            cpu:
              type: string
              description: CPU request (e.g., 50m, 0.1, 1)
              example: 50m
            memory:
              type: string
              description: Memory request (e.g., 64Mi, 128Mi, 1Gi)
              example: 64Mi
        recommended:
          type: object
          description: Recommended resources for typical production use
          properties:
            cpu:
              type: string
              example: 100m
            memory:
              type: string
              example: 128Mi
        high:
          type: object
          description: Resources for high-traffic or compute-intensive scenarios
          properties:
            cpu:
              type: string
              example: 500m
            memory:
              type: string
              example: 512Mi
...
    FilterOption:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          description: The data type of the filter option
          enum:
            - string
            - number
        values:
          type: array
          description: Known values of the property for string types with a small number of possible options.
          items: {}
        range:
          $ref: "#/components/schemas/FilterOptionRange"
    FilterOptionRange:
      type: object
      description: Min and max values for number types.
      properties:
        min:
          type: number
          format: double
        max:
          type: number
          format: double
    FieldFilter:
      type: object
      required:
        - operator
        - value
      properties:
        operator:
          type: string
          description: Filter operator (e.g., '<', '=', '>', 'IN')
          example: '<'
        value:
          description: Filter value (can be number, string, or array)
          example: 70
    FilterOptionsList:
      description: List of FilterOptions
      type: object
      properties:
        filters:
          type: object
          description: A single filter option.
          additionalProperties:
            $ref: "#/components/schemas/FilterOption"
        namedQueries:
          type: object
          description: Predefined named queries for common filtering scenarios
          additionalProperties:
            type: object
            additionalProperties:
              $ref: "#/components/schemas/FieldFilter"
    OrderByField:
      description: |-
        Supported fields for ordering result entities.
      enum:
        - CREATE_TIME
        - LAST_UPDATE_TIME
        - ID
        - NAME
      type: string
  responses:
...
    MCPServerListResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MCPServerList"
      description: A response containing a list of MCP server entities.
    MCPServerResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MCPServer"
      description: A response containing an `MCPServer` entity.
    MCPToolsListResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MCPToolsList"
      description: A response containing a list of MCP tool entities.
...
    MCPToolWithServerResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MCPToolWithServer"
      description: A response containing an `MCPToolWithServer` entity.
    FilterOptionsResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FilterOptionsList"
      description: A response containing options for a `filterQuery` parameter.
...
  parameters:
    mcpServerFilterQuery:
      examples:
        mcpServerFilterQuery:
          value: "license='Apache 2.0' AND securityIndicators.verifiedSource=true"
      name: filterQuery
      description: |
        A SQL-like query string to filter MCP servers.

        **Supported Operators:**
        - Comparison: `=`, `!=`, `<>`, `>`, `<`, `>=`, `<=`
        - Pattern matching: `LIKE`, `ILIKE` (case-insensitive)
        - Set membership: `IN`
        - Logical: `AND`, `OR`
        - Grouping: `()` for complex expressions

        **Examples:**
        - `license = "Apache 2.0"`
        - `verifiedSource = true`
        - `provider ILIKE "%local%"`
        - `(license = "Apache 2.0" OR license = "MIT") AND verifiedSource = true`
      schema:
        type: string
        pattern: "^[\\x20-\\x7E]*$"
      in: query
      required: false
    mcpToolFilterQuery:
      examples:
        mcpToolFilterQuery:
          value: "accessType='read_only' AND name ILIKE '%list%'"
      name: filterQuery
      description: |
        A SQL-like query string to filter MCP tools for a specific MCP server.

        **Supported Operators:**
        - Comparison: `=`, `!=`, `<>`, `>`, `<`, `>=`, `<=`
        - Pattern matching: `LIKE`, `ILIKE` (case-insensitive)
        - Set membership: `IN`
        - Logical: `AND`, `OR`
        - Grouping: `()` for complex expressions

        **Examples:**
        - `name = "list_models"`
        - `accessType = "read_only"`
        - `name ILIKE "%search%"`
        - `(accessType = "read_only" OR accessType = "execute") AND name LIKE "%model%"`
      schema:
        type: string
        pattern: "^[\\x20-\\x7E]*$"
      in: query
      required: false
...
    orderBy:
      style: form
      explode: true
      examples:
        orderBy:
          value: ID
      name: orderBy
      description: Specifies the order by criteria for listing entities.
      schema:
        $ref: "#/components/schemas/OrderByField"
      in: query
      required: false
...
    namedQuery:
      examples:
        namedQuery:
          value: secure_servers
      name: namedQuery
      description: Predefined filter template name to apply when listing MCP servers.
      schema:
        type: string
      in: query
      required: false
    includeTools:
      examples:
        includeTools:
          value: true
      name: includeTools
      description: Whether to include the tools array in each MCP server result.
      schema:
        type: boolean
        default: false
      in: query
      required: false
    toolLimit:
      examples:
        toolLimit:
          value: 10
      name: toolLimit
      description: Maximum number of tools to include when includeTools is true.
      schema:
        type: integer
        format: int32
        default: 10
        minimum: 0
        maximum: 100
      in: query
      required: false
...