# catalog/internal/server/openapi/api_mcp_catalog_service.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Catalog REST API
 *
 * REST API for Kubeflow Hub Catalog
 *
 * API version: v1alpha1
 */

package openapi

import (
	"net/http"
	"strings"

	"github.com/go-chi/chi/v5"

	model "github.com/kubeflow/hub/catalog/pkg/openapi"
)

// MCPCatalogServiceAPIController binds http requests to an api service and writes the service results to the http response
type MCPCatalogServiceAPIController struct {
	service      MCPCatalogServiceAPIServicer
	errorHandler ErrorHandler
}

// MCPCatalogServiceAPIOption for how the controller is set up.
type MCPCatalogServiceAPIOption func(*MCPCatalogServiceAPIController)

// WithMCPCatalogServiceAPIErrorHandler inject ErrorHandler into controller
func WithMCPCatalogServiceAPIErrorHandler(h ErrorHandler) MCPCatalogServiceAPIOption {
	return func(c *MCPCatalogServiceAPIController) {
		c.errorHandler = h
	}
}

// NewMCPCatalogServiceAPIController creates a default api controller
func NewMCPCatalogServiceAPIController(s MCPCatalogServiceAPIServicer, opts ...MCPCatalogServiceAPIOption) *MCPCatalogServiceAPIController {
	controller := &MCPCatalogServiceAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the MCPCatalogServiceAPIController
func (c *MCPCatalogServiceAPIController) Routes() Routes {
	return Routes{
		"FindMCPServers": Route{
			"FindMCPServers",
			strings.ToUpper("Get"),
			"/api/mcp_catalog/v1alpha1/mcp_servers",
			c.FindMCPServers,
		},
		"FindMCPServersFilterOptions": Route{
			"FindMCPServersFilterOptions",
			strings.ToUpper("Get"),
			"/api/mcp_catalog/v1alpha1/mcp_servers/filter_options",
			c.FindMCPServersFilterOptions,
		},
		"GetMCPServer": Route{
			"GetMCPServer",
			strings.ToUpper("Get"),
			"/api/mcp_catalog/v1alpha1/mcp_servers/{server_id}",
			c.GetMCPServer,
		},
		"FindMCPServerTools": Route{
			"FindMCPServerTools",
			strings.ToUpper("Get"),
			"/api/mcp_catalog/v1alpha1/mcp_servers/{server_id}/tools",
			c.FindMCPServerTools,
		},
		"GetMCPServerTool": Route{
			"GetMCPServerTool",
			strings.ToUpper("Get"),
			"/api/mcp_catalog/v1alpha1/mcp_servers/{server_id}/tools/{tool_name}",
			c.GetMCPServerTool,
		},
	}
}

// OrderedRoutes returns all the api routes in a deterministic order for the MCPCatalogServiceAPIController
func (c *MCPCatalogServiceAPIController) OrderedRoutes() []Route {
	return []Route{
		Route{
			"FindMCPServers",
			strings.ToUpper("Get"),
			"/api/mcp_catalog/v1alpha1/mcp_servers",
			c.FindMCPServers,
		},
		Route{
			"FindMCPServersFilterOptions",
			strings.ToUpper("Get"),
			"/api/mcp_catalog/v1alpha1/mcp_servers/filter_options",
			c.FindMCPServersFilterOptions,
		},
		Route{
			"GetMCPServer",
			strings.ToUpper("Get"),
			"/api/mcp_catalog/v1alpha1/mcp_servers/{server_id}",
			c.GetMCPServer,
		},
		Route{
			"FindMCPServerTools",
			strings.ToUpper("Get"),
			"/api/mcp_catalog/v1alpha1/mcp_servers/{server_id}/tools",
			c.FindMCPServerTools,
		},
		Route{
			"GetMCPServerTool",
			strings.ToUpper("Get"),
			"/api/mcp_catalog/v1alpha1/mcp_servers/{server_id}/tools/{tool_name}",
			c.GetMCPServerTool,
		},
	}
}

// FindMCPServers - List MCP servers.
func (c *MCPCatalogServiceAPIController) FindMCPServers(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	var nameParam string
	if query.Has("name") {
		param := query.Get("name")

		nameParam = param
	} else {
	}
	var qParam string
	if query.Has("q") {
		param := query.Get("q")

		qParam = param
	} else {
	}
	var sourceLabelParam []string
	if query.Has("sourceLabel") {
		sourceLabelParam = strings.Split(query.Get("sourceLabel"), ",")
	}
	var filterQueryParam string
	if query.Has("filterQuery") {
		param := query.Get("filterQuery")

		filterQueryParam = param
	} else {
	}
	var namedQueryParam string
	if query.Has("namedQuery") {
		param := query.Get("namedQuery")

		namedQueryParam = param
	} else {
	}
	var includeToolsParam bool
	if query.Has("includeTools") {
		param, err := parseBoolParameter(
			query.Get("includeTools"),
			WithParse[bool](parseBool),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "includeTools", Err: err}, nil)
			return
		}

		includeToolsParam = param
	} else {
		var param bool = false
		includeToolsParam = param
	}
	var toolLimitParam int32
	if query.Has("toolLimit") {
		param, err := parseNumericParameter[int32](
			query.Get("toolLimit"),
			WithParse[int32](parseInt32),
			WithMinimum[int32](0),
			WithMaximum[int32](100),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "toolLimit", Err: err}, nil)
			return
		}

		toolLimitParam = param
	} else {
		var param int32 = 10
		toolLimitParam = param
	}
	var pageSizeParam string
	if query.Has("pageSize") {
		param := query.Get("pageSize")

		pageSizeParam = param
	} else {
	}
	var orderByParam model.OrderByField
	if query.Has("orderBy") {
		param := model.OrderByField(query.Get("orderBy"))

		orderByParam = param
	} else {
	}
	var sortOrderParam model.SortOrder
	if query.Has("sortOrder") {
		param := model.SortOrder(query.Get("sortOrder"))

		sortOrderParam = param
	} else {
	}
	var nextPageTokenParam string
	if query.Has("nextPageToken") {
		param := query.Get("nextPageToken")

		nextPageTokenParam = param
	} else {
	}
	result, err := c.service.FindMCPServers(r.Context(), nameParam, qParam, sourceLabelParam, filterQueryParam, namedQueryParam, includeToolsParam, toolLimitParam, pageSizeParam, orderByParam, sortOrderParam, nextPageTokenParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// FindMCPServersFilterOptions - Lists fields, values, and named queries that can be used in `filterQuery` on the list MCP servers endpoint.
func (c *MCPCatalogServiceAPIController) FindMCPServersFilterOptions(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.FindMCPServersFilterOptions(r.Context())
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetMCPServer - Get an `MCPServer`.
func (c *MCPCatalogServiceAPIController) GetMCPServer(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	serverIdParam := chi.URLParam(r, "server_id")
	if serverIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"server_id"}, nil)
		return
	}
	var includeToolsParam bool
	if query.Has("includeTools") {
		param, err := parseBoolParameter(
			query.Get("includeTools"),
			WithParse[bool](parseBool),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "includeTools", Err: err}, nil)
			return
		}

		includeToolsParam = param
	} else {
		var param bool = false
		includeToolsParam = param
	}
	var toolLimitParam int32
	if query.Has("toolLimit") {
		param, err := parseNumericParameter[int32](
			query.Get("toolLimit"),
			WithParse[int32](parseInt32),
			WithMinimum[int32](0),
			WithMaximum[int32](100),
		)
		if err != nil {
			c.errorHandler(w, r, &ParsingError{Param: "toolLimit", Err: err}, nil)
			return
		}

		toolLimitParam = param
	} else {
		var param int32 = 10
		toolLimitParam = param
	}
	result, err := c.service.GetMCPServer(r.Context(), serverIdParam, includeToolsParam, toolLimitParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// FindMCPServerTools - List tools exposed by an `MCPServer`.
func (c *MCPCatalogServiceAPIController) FindMCPServerTools(w http.ResponseWriter, r *http.Request) {
	query, err := parseQuery(r.URL.RawQuery)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	serverIdParam := chi.URLParam(r, "server_id")
	if serverIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"server_id"}, nil)
		return
	}
	var filterQueryParam string
	if query.Has("filterQuery") {
		param := query.Get("filterQuery")

		filterQueryParam = param
	} else {
	}
	var pageSizeParam string
	if query.Has("pageSize") {
		param := query.Get("pageSize")

		pageSizeParam = param
	} else {
	}
	var orderByParam model.OrderByField
	if query.Has("orderBy") {
		param := model.OrderByField(query.Get("orderBy"))

		orderByParam = param
	} else {
	}
	var sortOrderParam model.SortOrder
	if query.Has("sortOrder") {
		param := model.SortOrder(query.Get("sortOrder"))

		sortOrderParam = param
	} else {
	}
	var nextPageTokenParam string
	if query.Has("nextPageToken") {
		param := query.Get("nextPageToken")

		nextPageTokenParam = param
	} else {
	}
	result, err := c.service.FindMCPServerTools(r.Context(), serverIdParam, filterQueryParam, pageSizeParam, orderByParam, sortOrderParam, nextPageTokenParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetMCPServerTool - Get an `MCPTool` from an `MCPServer`.
func (c *MCPCatalogServiceAPIController) GetMCPServerTool(w http.ResponseWriter, r *http.Request) {
	serverIdParam := chi.URLParam(r, "server_id")
	if serverIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"server_id"}, nil)
		return
	}
	toolNameParam := chi.URLParam(r, "tool_name")
	if toolNameParam == "" {
		c.errorHandler(w, r, &RequiredError{"tool_name"}, nil)
		return
	}
	result, err := c.service.GetMCPServerTool(r.Context(), serverIdParam, toolNameParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}